recipes-core: images: unset IMAGE_NAME_SUFFIX for QSPI NAND and NOR images
Commit 26d97ac of openembedded-core introduced a change to
image-artifact-names, adding ${IMAGE_NAME_SUFFIX} directly to both
${IMAGE_NAME} and ${IMAGE_LINK_NAME}. This change appended a 'rootfs'
suffix to the NAND and NOR image names, which deviates from the
previously defined naming convention for the .mtd images. Although this
change is cosmetic, setting IMAGE_NAME_SUFFIX to an empty string ("")
would allow us to retain the previous format, which aligns with the
naming convention of previously released assets and ensures
compatibility with the internal CI pipelines.

For more information please refer to the following commit:

https://github.com/openembedded/openembedded-core/commit/26d97acc71379ab6702fa54a23b6542a3f51779c

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>