recipes-core: systemd: drop local systemd v254.4
Drop the local version of systemd (v254.4) used in the previous release.
This is no longer needed because scrathgap 5.0.3 uses version 255.4
which supports repart and growfs.

Also, update the systemd-networkd-wait-online service patch to
the latest systemd version and add the "Upstream-Status" field in the patch
which is now required in scarthgap to avoid QA warnings/issues.

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>