recipes-core: images: fix IMAGE_INSTALL append in core-image-minimal-dev
The core-image-minimal-dev bbappend in the PIC64GX BSP layer is
supposed to append extra packages to the existing core-image-minimal-dev
recipe define in the OE-core layer. However, the IMAGE_INSTALL variable
has a "=" assignment instead of an append "+=" assignment, which is
redifining the original packges in the original core-iamge-minimal-dev
recipe. Fix the variable assignment from "=" to "+=".

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>
Signed-off-by: Pierre-Henry Moussay <pierre-henry.moussay@microchip.com>