recipes-kernel: pic64gx-linux: set KERNEL_DTBVENDORED bitbake variable
Commit 941ba1a in openembedded-core added a "KERNEL_DTBVENDORED" variable
which allows to select if the device tree subdirectory is added in the
node name of the FITImage.

In previous releases of PIC64GX SoC, the "microchip" keyword was added
to the conf section of the FITImage.

Set the KERNEL_DTBVENDORED to true to keep the same FIT image node
names as in previous releases for compatibility purposes.

For more information please refer to commit 941ba1a:
https://github.com/openembedded/openembedded-core/commit/941ba1a132bafa9c9be855fb91fec96d8b06299f

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>
Signed-off-by: Pierre-Henry Moussay <pierre-henry.moussay@microchip.com>