recipes-bsp, recipes-kernel: set bootargs in u-boot env and use earlycon
Set bootargs in U-Boot in order to align with what is done with Polarfire
Yocto distribution.

Also, the earlycon=sbi argument previously passed to the kernel is no
longer applicable, as it requires the SBI 0.1.

earlycon is the recommended bootline argument to use.
The UARTs in the MSS are compatible with the ns16550a standard, and
there is already an earlycon implementation for those. As a result,
the kernel can directly access the UART without going through the SBI.

Signed-off-by: Valentina Fernandez <valentina.fernandezalanis@microchip.com>
Signed-off-by: Pierre-Henry Moussay <pierre-henry.moussay@microchip.com>