pic64gx-zephyr-amp-demo: update zephyr demo
The demo is now based on OpenAMP, 2 firmware must be compiled.
One for the HSS boot because Linux needs a pre-defined vdev address to
be able to attach to the already running firmware. For the standard use
the resource table location is just extracted from elf before loading, but
doesn't support well loaded section.

Signed-off-by: Pierre-Henry Moussay <pierre-henry.moussay@microchip.com>